27. Remove Element
